+++ lib/russianpost/barcode/invalid.rb
@@ -1,0 +1,11 @@
+require 'russianpost/barcode/base'
+
+module RussianPost
+  module Barcode
+    class Invalid < Base
+      def valid?
+        false
+      end
+    end
+  end
+end
